JobSystemにワーカークラスを追加し、JobManager改修
