+++ entityProject.vcxproj.filters
@@ -134,5 +134,14 @@
     <ClInclude Include="intrusive_ptr.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="TaskQueue.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ChunkAllocator.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="JobWorker.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
